Update a line of code in Parallax

+++ Assets/Scripts/Parallax.cs
@@ -15,7 +15,7 @@ public class Parallax : MonoBehaviour {
             if (child.position.y >= endY) {
                 child.position = new Vector3(child.position.x, endY * -1, child.position.z); 
             }
-            backgroundSpeed = backgroundSpeed / (float)1.5; 
+            backgroundSpeed = backgroundSpeed / 1.5f; 
         }
     }
 }
